Trabajo terminado
Se agregan validaciones en el ABM (Insert, Update, Delete)

+++ Lab.Practica3.EF/Lab.Practica3.EF.Logic/EmployeeLogic.cs
@@ -18,24 +18,37 @@ namespace Lab.Practica6.Logic
 
         public void Delete(Employee entity)
         {
+            if (entity.EmployeeID == 0)
+            {
+                throw new Exception("Los campos no pueden ser nulos");
+            }
+
             Employee employeeToDelete = context.Employee.Find(entity.EmployeeID);
            
             context.Employee.Remove(employeeToDelete);
 
             context.SaveChanges();                                     
         }
+
         public void Insert(Employee entity)
         {
             if (entity.FirstName == null || entity.LastName == null)
             {
                 throw new Exception("Los campos no pueden ser nulos");
             }
+
             context.Employee.Add(entity);
    
             context.SaveChanges();
         }
+
         public void Update(Employee entity)
-        {              
+        {
+            if (entity.FirstName == null || entity.LastName == null)
+            {
+                throw new Exception("Los campos no pueden ser nulos");
+            }
+
             var employeeUpdate = context.Employee.Find(entity.EmployeeID);
 
             employeeUpdate.LastName = entity.LastName;
